+++ src/main/java/com/tecfrac/helpdesk/repository/TicketTypeRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TicketTypeRepository extends JpaRepository<ModelTicketType, Integer> {
+public interface TicketTypeRepository extends JpaRepository<ModelTicketType, Long> {
 
     /**
      *
